Move ScanValidator to protocol namespace
